+++ package/layout/PageContainer.js
@@ -61,9 +61,10 @@ export default function PageContainer(props) {
   validationsRegistry = React.useContext(ValidationsRegistryContext);
 
   // -- console.log("mergedComponentRegistry", mergedComponentRegistry, mergedResourceRegistry);
-  const { user: { id: userID }, accessToken, sessionExpired, sessionDetail } = useSelector((state) => state?.auth || {});
-  
-  let authenticated = accessToken ? true : false;
+  const authReducerValues =  useSelector((state) => state?.auth || {});
+  const userID = authReducerValues?.user?.id || {};
+  const { accessToken, sessionExpired, sessionDetail } = authReducerValues || {};
+  let authenticated = !!accessToken;
 
   const snackMessages = useSelector((state) => state?.app?.snackMessages || []);
   const appState = useSelector((state) => state?.app || []);