+++ package/route/CoreRoutesProvider.js
@@ -1,29 +1,36 @@
-
 import React from "react";
 
 // eslint-disable-next-line import/no-unresolved
+import { WrappidDataContext } from "@wrappid/styles";
 import { useDispatch, useSelector } from "react-redux";
 
 import { HTTP } from "../config/constants";
 import { CoreRoutesContext } from "../config/contextHandler";
 import { RoutesRegistry } from "../registry/RoutesRegistry";
 import { apiRequestAction } from "../store/action/appActions";
-import { GET_ROUTE_FAILURE, GET_ROUTE_SUCCESS, LOCAL_ROUTES_SYNCED_SUCCESS } from "../store/types/appTypes";
+import {
+  GET_ROUTE_FAILURE,
+  GET_ROUTE_SUCCESS,
+  LOCAL_ROUTES_SYNCED_SUCCESS
+} from "../store/types/appTypes";
 
 export default function CoreRoutesProvider(props) {
+  const { config } = React.useContext(WrappidDataContext);
   const { routes: localRoutes = {}, children } = props;
   const dispatch = useDispatch();
-  const { sync, routes: storedRoutes = {} } = useSelector((state) => state?.route);
+  const { sync, routes: storedRoutes = {} } = useSelector(
+    (state) => state?.route
+  );
   const { local: localSync = false, server: serverSync = false } = sync;
   const { accessToken } = useSelector((state) => state?.auth || {});
-  let authenticated = accessToken ? true : false;
-
+  let authenticated = !!accessToken;
   const [contextSync, setContextSync] = React.useState(false);
   const [routesList, setRoutesList] = React.useState({});
+  const { dynamic: isDynamic } = config || {};
 
   React.useEffect(() => {
     // @todo post sync
-    if (!serverSync) {
+    if (!serverSync && isDynamic) {
       dispatch(
         apiRequestAction(
           HTTP.GET,
@@ -40,27 +47,37 @@ export default function CoreRoutesProvider(props) {
   React.useEffect(() => {
     /**
      * @todo call DATA_SYNC_API
-     * 
-    */
+     *
+     */
     if (!localSync && Object.keys(localRoutes)?.length > 0) {
-      dispatch({ payload: { ...RoutesRegistry, ...localRoutes }, type: LOCAL_ROUTES_SYNCED_SUCCESS });
+      dispatch({
+        payload: { ...RoutesRegistry, ...localRoutes },
+        type   : LOCAL_ROUTES_SYNCED_SUCCESS,
+      });
     }
   }, [localSync, localRoutes]);
 
   React.useEffect(() => {
     let _storedRoutes = {};
-    
-    if (localSync && serverSync && !contextSync) {
-      setContextSync(true);
 
-      storedRoutes.forEach(route => {
+    if ((localSync || serverSync) && !contextSync) {
+      setContextSync(true);
+      storedRoutes.forEach((route) => {
         _storedRoutes[route?.entityRef] = route;
       });
-  
       setRoutesList(_storedRoutes);
     }
   }, [storedRoutes, localSync, serverSync, contextSync]);
-    
+
+  // effect to set contextSync to false after API call
+  // if serverSync changes from false to true, API call has completed.
+  React.useEffect(() => {
+    if (serverSync) {
+      // reset contextSync to allow re-sync after API call completion.
+      setContextSync(false);
+    }
+  }, [serverSync]);
+
   return (
     <CoreRoutesContext.Provider value={routesList}>
       {children}